requirements.txt
# local package
-e .

# external requirements
click~=8.0.3
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1

# pipreqs outputs minus those mentioned above already
hydra-core==1.1.1
omegaconf==2.1.1
Pillow~=8.4.0
pytorch_lightning==1.5.8
requests==2.26.0
setuptools~=60.3.1
torch==1.10.1
torchvision==0.11.2
transformers==4.15.0
